chat_service/requirements.txt
Django==4.2
psycopg2-binary==2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.13.0
requests
celery
redis
pika
daphne==4.1.2
channels==4.1.0
channels-redis
aiohttp==3.9.5